requirements.txt
requests
scratchattach
python-dotenv